Move solvers out of fmi2 namespace and rename classes
