+++ include/fmi4cpp/fmi2/fmi4cpp.hpp
@@ -36,8 +36,4 @@
 #include "import/CoSimulationSlave.hpp"
 #include "import/ModelExchangeSlave.hpp"
 
-#if FMI4CPP_WITH_ODEINT
-#include "import/OdeintSolver.hpp"
-#endif
-
 #endif //FMI4CPP_FMI4CPP_FMI2_HPP
