+++ include/fmi4cpp/fmi2/import/Fmu.hpp
@@ -146,7 +146,7 @@ namespace fmi4cpp::fmi2 {
         std::unique_ptr<ModelExchangeInstance> newInstance(bool visible = false, bool loggingOn = false);
 
         std::unique_ptr<ModelExchangeSlave>
-        newInstance(std::unique_ptr<Solver> &solver, bool visible = false, bool loggingOn = false);
+        newInstance(std::unique_ptr<solver::ModelExchangeSolver> &solver, bool visible = false, bool loggingOn = false);
 
     };
